requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mimerender==0.6.0
nose==1.3.7
psycopg2==2.7.3.2
pyodbc==4.0.22
python-dateutil==2.6.1
python-mimeparse==1.6.0
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
Werkzeug==0.14.1
coverage==4.4.2
